Restructuring payload extensions
